Moved bass drum pedal logic to its own app state
